componetiza modal de erro
